etablishment ray horizontal

+++ src/cast_All_Rays/distanceBetweenPoints.c
@@ -1,7 +1,6 @@
-#include "cub3d.h"
+#include <math.h>
 
-float	distanceBetweenPoints(int x1, int y1, int x2, int y2)
+float	distanceBetweenPoints(float x1, float y1, float x2, float y2)
 {
 	return (sqrt((x2 - x1) * (x2 - x1) + (y2 - y1) * (y2 - y1)));
 }
-
